Move Rust API unit tests to a separate directory
